Modifications for building under VC. - Added missing C99 functions.
git-svn-id: https://svn.code.sf.net/p/freemat/code/branches/cmake_build@2904 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ libs/libFreeMat/Print.cpp
@@ -180,13 +180,13 @@ void ComputeScaleFactor(const T* array, int count, ArrayFormatInfo& format) {
   }
   if (!finiteElementFound) return;
   if (max_amplitude >= 100)
-    format.scalefact = pow(10.0,floor(log10(max_amplitude)));
+    format.scalefact = pow(10.0,static_cast<double>(floor(log10(max_amplitude))));
   else if (max_amplitude <= -100)
-    format.scalefact = pow(10.0,floor(log10(-max_amplitude)));
+    format.scalefact = pow(10.0,static_cast<double>(floor(log10(-max_amplitude))));
   else if ((max_amplitude <= .1) && (max_amplitude>0))
-    format.scalefact = pow(10.0,floor(log10(max_amplitude)));
+    format.scalefact = pow(10.0,static_cast<double>(floor(log10(max_amplitude))));
   else if ((max_amplitude >= -.1) && (max_amplitude<0))
-    format.scalefact = pow(10.0,floor(log10(-max_amplitude)));
+    format.scalefact = pow(10.0,static_cast<double>(floor(log10(-max_amplitude))));
   else
     format.scalefact = 1.0;
 }
